Added EC2 support for on-demand instances, since spot instances appear very volatile due to price spikes.

+++ aws_opacmo.sh
@@ -60,6 +60,28 @@ wait_for_completion() {
 	echo ''
 }
 
+help_message() {
+	echo "Usage: aws_opacmo.sh instance_type"
+	echo ""
+	echo "Note: avoid using this script on its own. Use make_opacmo.sh instead."
+	echo ""
+	echo "Options for 'instance_type':"
+	echo "  ondemand : use on-demand instances (fixed price, guaranteed availability once started)"
+	echo "  spot     : use spot instances (variable price, availability not guaranteed)"
+}
+
+if [[ ! $# -eq 1 ]] ; then
+	help_message
+	exit 2
+fi
+
+if [ "$1" != 'ondemand' ] && [ "$1" != 'spot' ] ; then
+	help_message
+	exit 2
+fi
+
+instance_type=$1
+
 # Create a tmp directory, if it not already exists. This will be used to store
 # transient data, s.a. AWS pricing and computer generated scripts that are executed
 # on the "worker" instances:
@@ -67,42 +89,44 @@ if [ ! -d tmp ] ; then
 	mkdir tmp
 fi
 
-# Determine suitable price:
-N=0
-FACTOR=1.5
-AVG_PRICE=0.0
-ec2-describe-spot-price-history -t $instance_type --product-description 'Linux/UNIX' | cut -f 2 -d '	' | sort -n > tmp/aws_prices.tmp
-for price in ` cat tmp/aws_prices.tmp` ; do
-	AVG_PRICE=`echo "$AVG_PRICE+$price" | bc`
-	let N=N+1
-done
-MEDIAN_PRICE=`$awk_interpreter '{ count[NR] = $1; } END { if (NR % 2) { print count[(NR + 1) / 2]; } else { print (count[(NR / 2)] + count[(NR / 2) + 1]) / 2.0; } }' tmp/aws_prices.tmp`
-MEDIAN_PRICE=`echo "scale=3;$MEDIAN_PRICE/1" | bc | sed $sed_regexp 's/^\./0./'`
-AVG_PRICE=`echo "scale=3;$AVG_PRICE/$N" | bc | sed $sed_regexp 's/^\./0./'`
-MAX_PRICE=`echo "scale=3;$MEDIAN_PRICE*$FACTOR" | bc | sed $sed_regexp 's/^\./0./'`
-rm -f tmp/aws_prices.tmp
+if [ "$instance_type" = 'spot' ] ; then
+	# Determine suitable price:
+	N=0
+	FACTOR=1.5
+	AVG_PRICE=0.0
+	ec2-describe-spot-price-history -t $instance_type --product-description 'Linux/UNIX' | cut -f 2 -d '	' | sort -n > tmp/aws_prices.tmp
+	for price in ` cat tmp/aws_prices.tmp` ; do
+		AVG_PRICE=`echo "$AVG_PRICE+$price" | bc`
+		let N=N+1
+	done
+	MEDIAN_PRICE=`$awk_interpreter '{ count[NR] = $1; } END { if (NR % 2) { print count[(NR + 1) / 2]; } else { print (count[(NR / 2)] + count[(NR / 2) + 1]) / 2.0; } }' tmp/aws_prices.tmp`
+	MEDIAN_PRICE=`echo "scale=3;$MEDIAN_PRICE/1" | bc | sed $sed_regexp 's/^\./0./'`
+	AVG_PRICE=`echo "scale=3;$AVG_PRICE/$N" | bc | sed $sed_regexp 's/^\./0./'`
+	MAX_PRICE=`echo "scale=3;$MEDIAN_PRICE*$FACTOR" | bc | sed $sed_regexp 's/^\./0./'`
+	rm -f tmp/aws_prices.tmp
 
-echo "Over $N reported prices, all zones (via 'ec2-describe-spot-price-history'):"
-echo "Average price: $AVG_PRICE"
-echo "Median price: $MEDIAN_PRICE"
-echo ""
-echo "Suggest max. price for opacmo run: $MAX_PRICE (${FACTOR}x median price)"
+	echo "Over $N reported prices, all zones (via 'ec2-describe-spot-price-history'):"
+	echo "Average price: $AVG_PRICE"
+	echo "Median price: $MEDIAN_PRICE"
+	echo ""
+	echo "Suggest max. price for opacmo run: $MAX_PRICE (${FACTOR}x median price)"
 
-echo -n "Type 'yes' (without the quotes) to accept, or enter a max. price (e.g., 0.70): "
-read user_agreement_or_price
+	echo -n "Type 'yes' (without the quotes) to accept, or enter a max. price (e.g., 0.70): "
+	read user_agreement_or_price
 
-if [ "$user_agreement_or_price" != 'yes' ] ; then
-	if [ "`echo -n "$user_agreement_or_price" | grep -o -E '^[0-9]+\.[0-9]+$'`" != "$user_agreement_or_price\n" ] ; then
-		MAX_PRICE=$user_agreement_or_price
-		echo -n "Type 'yes' to accept your custom price of $MAX_PRICE and continue: "
-		read user_agreement
-		if [ "$user_agreement" != 'yes' ] ; then
-			echo 'You declined your suggested price. Aborting.'
-			exit 2
+	if [ "$user_agreement_or_price" != 'yes' ] ; then
+		if [ "`echo -n "$user_agreement_or_price" | grep -o -E '^[0-9]+\.[0-9]+$'`" != "$user_agreement_or_price\n" ] ; then
+			MAX_PRICE=$user_agreement_or_price
+			echo -n "Type 'yes' to accept your custom price of $MAX_PRICE and continue: "
+			read user_agreement
+			if [ "$user_agreement" != 'yes' ] ; then
+				echo 'You declined your suggested price. Aborting.'
+				exit 3
+			fi
+		else
+			echo 'You declined the suggested price. Aborting.'
+			exit 4
 		fi
-	else
-		echo 'You declined the suggested price. Aborting.'
-		exit 3
 	fi
 fi
 
@@ -116,25 +140,32 @@ echo "Setting up 'opacmo_$TIMESTAMP' security group..."
 SECURITY_GROUP=`ec2-create-group --description 'opacmo security group' opacmo_$TIMESTAMP | cut -f 2 -d '	'`
 if [ "$SECURITY_GROUP" = '' ] ; then
 	echo "Could not create the security group 'opacmo' (via ec2-create-group). Does it already exist?"
-	exit 4
+	exit 5
 fi
 echo "Security group 'opacmo_$TIMESTAMP' created: $SECURITY_GROUP"
 ec2-authorize $SECURITY_GROUP -p 22
 ec2-authorize $SECURITY_GROUP -p 80
 ec2-authorize $SECURITY_GROUP -o $SECURITY_GROUP -u $AWS_ACCOUNT_ID
 
-echo "Requesting spot instance (via ec2-request-spot-instances)..."
-SPOT_INSTANCE_REQUEST=`ec2-request-spot-instances -g opacmo_$TIMESTAMP -p $MAX_PRICE -k $AWS_KEY_PAIR -z $zone -t $instance_type -b '/dev/sda2=ephemeral0' --user-data-file opacmo/ec2/cache.sh $ami | cut -f 2 -d '	'`
-echo "Spot instance request filed: $SPOT_INSTANCE_REQUEST"
+if [ "$instance_type" = 'spot' ] ; then
+	echo "Requesting spot instance (via ec2-request-spot-instances)..."
+	SPOT_INSTANCE_REQUEST=`ec2-request-spot-instances -g opacmo_$TIMESTAMP -p $MAX_PRICE -k $AWS_KEY_PAIR -z $zone -t $instance_type -b '/dev/sda2=ephemeral0' --user-data-file opacmo/ec2/cache.sh $ami | cut -f 2 -d '	'`
+	echo "Spot instance request filed: $SPOT_INSTANCE_REQUEST"
 
-echo -n "Waiting for instance to boot."
-INSTANCE=
-while [ "$INSTANCE" = '' ] ; do
-	echo -n '.'
-	sleep $SPOT_CHECK_INTERVAL
-	INSTANCE=`ec2-describe-spot-instance-requests $SPOT_INSTANCE_REQUEST | cut -f 12 -d '	'`
-done
-echo ''
+	echo -n "Waiting for instance to boot."
+	INSTANCE=
+	while [ "$INSTANCE" = '' ] ; do
+		echo -n '.'
+		sleep $SPOT_CHECK_INTERVAL
+		INSTANCE=`ec2-describe-spot-instance-requests $SPOT_INSTANCE_REQUEST | cut -f 12 -d '	'`
+	done
+	echo ''
+fi
+
+if [ "$instance_type" = 'ondemand' ] ; then
+	echo "Requesting on-demand instance (via ec2-run-instances)..."
+	INSTANCE=`ec2-run-instances --instance-initiated-shutdown-behavior terminate -g opacmo_$TIMESTAMP -k $AWS_KEY_PAIR -z $zone -t $instance_type -b '/dev/sda2=ephemeral0' --user-data-file opacmo/ec2/cache.sh $ami | cut -f 2 -d '	'`
+fi
 
 echo "Instance started: $INSTANCE"
 echo -n "Getting IP addresses."
@@ -166,6 +197,13 @@ echo "Starting worker instances..."
 for prefix in ${WORKERS[@]} ; do
 	echo "Starting worker for journal prefix: $prefix"
 	sed $sed_regexp "s/PREFIX_VAR/$prefix/g" opacmo/ec2/worker.sh | sed $sed_regexp "s/CACHE_IP_VAR/$PRIVATE_CACHE_IP/g" > tmp/worker_$prefix.sh
-	ec2-request-spot-instances -g opacmo_$TIMESTAMP -p $MAX_PRICE -k $AWS_KEY_PAIR -z $zone -t $instance_type -b '/dev/sda2=ephemeral0' --user-data-file tmp/worker_$prefix.sh $ami
+
+	if [ "$instance_type" = 'spot' ] ; then
+		ec2-request-spot-instances -g opacmo_$TIMESTAMP -p $MAX_PRICE -k $AWS_KEY_PAIR -z $zone -t $instance_type -b '/dev/sda2=ephemeral0' --user-data-file tmp/worker_$prefix.sh $ami
+	fi
+
+	if [ "$instance_type" = 'ondemand' ] ; then
+		ec2-run-instances --instance-initiated-shutdown-behavior terminate -g opacmo_$TIMESTAMP -k $AWS_KEY_PAIR -z $zone -t $instance_type -b '/dev/sda2=ephemeral0' --user-data-file tmp/worker_$prefix.sh $ami
+	fi
 done
